+++ pages/all-benefits.js
@@ -6,26 +6,11 @@ import { withI18next } from "../lib/withI18next";
 import Layout from "../components/layout";
 import { connect } from "react-redux";
 import BenefitCard from "../components/benefit_cards";
+import { redux2i18n } from "../utils/redux2i18n";
 
 export class AllBenefits extends Component {
   componentWillMount() {
-    let i18nEn = {};
-    let i18nFr = {};
-    this.props.text.forEach(text => {
-      if (text.section) {
-        if (!i18nEn[text.section]) {
-          i18nEn[text.section] = {};
-          i18nFr[text.section] = {};
-        }
-        i18nEn[text.section][text.key] = text.English;
-        i18nFr[text.section][text.key] = text.French;
-      } else {
-        i18nEn[text.key] = text.English;
-        i18nFr[text.key] = text.French;
-      }
-    });
-    this.props.i18n.addResourceBundle("en", "common", i18nEn);
-    this.props.i18n.addResourceBundle("fr", "common", i18nFr);
+    redux2i18n(this.props.i18n, this.props.text);
   }
 
   render() {
